test: Cover CET <> CEST boundaries
